% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/script.R
\name{calArea}
\alias{calArea}
\title{Surface Area Calculator}
\usage{
calArea(poligono, unidad)
}
\arguments{
\item{poligono}{un objeto espacial sf}

\item{unidad}{unidad de medida de superficie en la que se quiera obtener el area (ha, m2)}
}
\value{
El area del poligono en la unidad de medida especificada
}
\description{
\code{calArea} calcula el area de un poligono en metros cuadrados o hectareas
}
\examples{
poligono <- sf::st_polygon(list(cbind(c(6, 10, 10, 6, 6), c(6, 6, 10, 10, 6))))
calArea(poligono, 'ha')
poligono <- sf::st_polygon(list(cbind(c(6, 10, 10, 6, 6), c(6, 6, 10, 10, 6))))
calArea(poligono, 'm2')

}
